fix: se modularizo la lista de usuarios y el inconveniente de redenrizado de la lista de usuarios
