+++ Cliente/VistaAgregar/AgregarControlador.js
@@ -1,8 +1,11 @@
 import { agregar } from "./mostrarAgregar.js"
+import { listarUsuario } from "../Peticiones/ListarUsuario.js"
 
 export const agregarControlador = () => {
  
     const editable = document.querySelector('div > div');
-    editable.innerHTML = agregar();
+
+    listarUsuario();
+
     
 }
