+++ Cliente/VistaEliminar/eliminarControlador.js
@@ -1,8 +1,10 @@
 import { eliminar } from "./mostrarEliminar.js"
+import { listarUsuario } from "../Peticiones/ListarUsuario.js"
 
 export const eliminarControlador = () => {
  
     const editable = document.querySelector('div > div');
-    editable.innerHTML = eliminar();
+    listarUsuario();
+
     
 }
